- Change rating dataType to float
- Debuging Usermapper tests, fix a bug in insert and acknowledged potential bug obtaining the current UnitOfWork using ThreadLocal at UnitOfWork.class

+++ SERVER/WEB_API/src/main/java/isel/ps/EmployBox/dal/dataMapping/mappers/CompanyMapper.java
@@ -28,7 +28,7 @@ public class CompanyMapper extends AccountMapper<Company> {
         try {
             long accountId = rs.getLong("accountId");
             String email = rs.getString("Email");
-            Double rating = rs.getDouble("Rating");
+            Float rating = rs.getFloat("Rating");
             String name = rs.getString("name");
             short yearFounded = rs.getShort("yearFounded");
             String specialization = rs.getString("Specialization");
